requirements.txt
# Core dependencies
requests>=2.31.0
dnspython>=2.4.0
python-whois>=0.8.0
phonenumbers>=8.13.0

# GUI dependencies
matplotlib>=3.7.0
networkx>=3.1

# PDF export
reportlab>=4.0.0

# Additional utilities
python-dateutil>=2.8.2
